Update to go-ipld-prime@v0.4.0
Updated the gen/main.go systems to new codegen;
they compile, but I have not exercised them.
IIUC, reapplying codegen in this package may require additional
manual intervention afterwards.  My goal in this diff is to move
forward the version of go-ipld-prime so that transitive dependencies
on passing through this library can be happier; it is not necessary
for me to reapply codegen today in order to accomplish this.